intra-range-methods {XVector}	R Documentation

Intra range transformations of an XVectorList object

Description

The intra range transformations are a set of generic functions defined in the IRanges package. Only 2 of them have methods for XVectorList objects: narrow and threebands. This man page describes those 2 methods only.

See ?`inter-range-methods` for more information.

Usage

## S4 method for signature 'XVectorList'
narrow(x, start=NA, end=NA, width=NA, use.names=TRUE)

## S4 method for signature 'XVectorList'
threebands(x, start=NA, end=NA, width=NA)

Arguments

x
An XVectorList object.

start, end, width
Vectors of integers, possibly with NAs. See the SEW (Start/End/Width) interface in the IRanges package for the details (?solveUserSEW).

use.names
TRUE or FALSE. Should names be preserved?

Details

narrow is equivalent to subset on an XVectorList object.

threebands extends the capability of narrow by returning the 3 XVectorList objects associated with the narrowing operation. The returned value y is a list of 3 XVectorList objects named "left", "middle" and "right". The middle component is obtained by calling narrow with the same arguments (except that names are dropped). The left and right components are also instances of the same class as x and they contain what has been removed on the left and right sides (respectively) of the original ranges during the narrowing.

Author(s)

H. Pagès

See Also

intra-range-methods in the IRanges package for intra range transformations.

solveUserSEW in the IRanges package for the SEW (Start/End/Width) interface.

The XVectorList class.

Examples

## ---------------------------------------------------------------------
## narrow()
## ---------------------------------------------------------------------

#TODO: show examples

## ---------------------------------------------------------------------
## threebands()
## ---------------------------------------------------------------------

#TODO: show examples

[Package XVector version 0.22.0 Index]