XRawList-comparison {XVector}	R Documentation

Comparing and ordering the list elements of XRawList objects

Description

Methods for comparing and ordering the elements in one or more XRawList objects.

Usage

## Element-wise (aka "parallel") comparison of 2 XRawList objects
## --------------------------------------------------------------

## S4 method for signature 'XRawList,XRawList'
e1 == e2

## S4 method for signature 'XRawList,XRawList'
e1 <= e2

## duplicated()
## ------------

## S4 method for signature 'XRawList'
duplicated(x, incomparables=FALSE, ...)

## match()
## -------

## S4 method for signature 'XRawList,XRawList'
match(x, table, nomatch=NA_integer_, incomparables=NULL)

## order() and related methods
## ---------------------------

## S4 method for signature 'XRawList'
is.unsorted(x, na.rm=FALSE, strictly=FALSE)

## S4 method for signature 'XRawList'
order(..., na.last=TRUE, decreasing=FALSE,
           method=c("auto", "shell", "radix"))

## S4 method for signature 'XRawList'
rank(x, na.last=TRUE,
     ties.method=c("average", "first", "random", "max", "min"))

## Generalized element-wise (aka "parallel") comparison of 2 XRawList objects
## --------------------------------------------------------------------------

## S4 method for signature 'XRawList,XRawList'
pcompare(x, y)

Arguments

e1, e2, x, table, y
XRawList objects.

incomparables
Not supported.

...
For duplicated: currently no additional arguments are allowed.

For order: additional XRawList objects used for breaking ties.

nomatch
The value to be returned in the case when no match is found. It is coerced to an integer.

na.rm, na.last, method
Ignored.

strictly
TRUE or FALSE. Should the check be for strictly increasing values?

decreasing
TRUE or FALSE.

ties.method
A character string specifying how ties are treated. Only "first" and "min" are supported for now.

Details

[TODO]

Author(s)

H. Pagès

See Also

The XRawList class.

Ranges-comparison in the IRanges package for comparing and ordering ranges.

==, duplicated, unique, match, %in%, order, sort, rank for general information about those functions.

Examples

## TODO

[Package XVector version 0.50.0 Index]